+++ workloads/inference/LLMs/offline-inference/vllm-batch-single-multinode/kaiwojob-llama-3.1-8b-instruct.yaml
@@ -21,13 +21,13 @@ spec:
     storageEnabled: true
     storageClassName: multinode
     data:
-      storageSize: 10Mi
+      storageSize: 20Mi
       mountPath: /workload
       download:
         git:
         - repository: https://github.com/silogen/kaiwo.git
           path: workloads/inference/LLMs/offline-inference/vllm-batch-single-multinode
-          branch: rocm6.4
+          branch: "rocm6.4"
           targetPath: code
     huggingFace:
       storageSize: "100Gi"
