+++ workloads/inference/LLMs/online-inference/vllm-online-single-multinode/__init__.py
@@ -14,9 +14,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+
 import logging
 import os
-from typing import List, Optional
+from typing import Dict, List, Optional
 
 from fastapi import FastAPI
 from ray import serve
@@ -31,7 +32,7 @@ from vllm.entrypoints.openai.protocol import (
     ErrorResponse,
 )
 from vllm.entrypoints.openai.serving_chat import OpenAIServingChat
-from vllm.entrypoints.openai.serving_engine import BaseModelPath, LoRAModulePath, PromptAdapterPath
+from vllm.entrypoints.openai.serving_engine import LoRAModulePath, PromptAdapterPath
 
 logger = logging.getLogger("ray.serve")
 
@@ -64,13 +65,22 @@ class VLLMDeployment:
 
     @app.post("/v1/chat/completions")
     async def create_chat_completion(self, request: ChatCompletionRequest, raw_request: Request):
+        """OpenAI-compatible HTTP endpoint.
+
+        API reference:
+            - https://docs.vllm.ai/en/latest/serving/openai_compatible_server.html
+        """
         if not self.openai_serving_chat:
             model_config = await self.engine.get_model_config()
-            base_model_paths = [BaseModelPath(name=self.engine_args.model, model_path=self.engine_args.model)]
+            # Determine the name of the served model for the OpenAI client.
+            if self.engine_args.served_model_name is not None:
+                served_model_names = self.engine_args.served_model_name
+            else:
+                served_model_names = [self.engine_args.model]
             self.openai_serving_chat = OpenAIServingChat(
                 self.engine,
                 model_config,
-                base_model_paths,
+                served_model_names,
                 self.response_role,
                 lora_modules=self.lora_modules,
                 prompt_adapters=self.prompt_adapters,
@@ -89,30 +99,48 @@ class VLLMDeployment:
             return JSONResponse(content=generator.model_dump())
 
 
-engine_args = AsyncEngineArgs(
-    model=os.getenv("MODEL_ID", "meta-llama/Meta-Llama-3-8B-Instruct"),
-    disable_log_requests=True,
-    dtype="auto",
-    device="cuda",
-    enable_chunked_prefill=False,
-    enable_prefix_caching=True,
-    gpu_memory_utilization=float(os.getenv("GPU_MEMORY_UTILIZATION", "0.8")),
-    max_model_len=int(os.getenv("MAX_MODEL_LEN", "4096")),
-    max_num_batched_tokens=int(os.getenv("MAX_NUM_BATCHED_TOKENS", "32768")),
-    max_num_seqs=int(os.getenv("MAX_NUM_SEQ", "512")),
-    pipeline_parallel_size=int(os.getenv("NUM_REPLICAS", "2")),
-    tensor_parallel_size=int(os.getenv("NUM_GPUS_PER_REPLICA", "8")),
-    tokenizer_pool_size=4,
-    tokenizer_pool_type="ray",
-    distributed_executor_backend="ray",
-    trust_remote_code=True,
-)
+def build_app(cli_args: Dict[str, str]) -> serve.Application:
+    """Builds the Serve app based on CLI arguments.
 
-deployment = VLLMDeployment.options().bind(
-    engine_args=engine_args,
-    response_role="assistant",
-    lora_modules=None,
-    prompt_adapters=None,
-    request_logger=None,
-    chat_template=None,
+    See https://docs.vllm.ai/en/latest/serving/openai_compatible_server.html#command-line-arguments-for-the-server
+    for the complete set of arguments.
+
+    Supported engine arguments: https://docs.vllm.ai/en/latest/models/engine_args.html.
+    """  # noqa: E501
+    engine_args = AsyncEngineArgs(
+        model=os.getenv("MODEL_ID", "meta-llama/Meta-Llama-3-8B-Instruct"),
+        disable_log_requests=True,
+        dtype="auto",
+        device="cuda",
+        enable_chunked_prefill=False,
+        enable_prefix_caching=True,
+        gpu_memory_utilization=float(os.getenv("GPU_MEMORY_UTILIZATION", "0.8")),
+        max_model_len=int(os.getenv("MAX_MODEL_LEN", "4096")),
+        max_num_batched_tokens=int(os.getenv("MAX_NUM_BATCHED_TOKENS", "32768")),
+        max_num_seqs=int(os.getenv("MAX_NUM_SEQ", "512")),
+        pipeline_parallel_size=int(os.getenv("NUM_REPLICAS", "2")),
+        tensor_parallel_size=int(os.getenv("NUM_GPUS_PER_REPLICA", "8")),
+        tokenizer_pool_size=4,
+        tokenizer_pool_type="ray",
+        distributed_executor_backend="ray",
+        trust_remote_code=True,
+    )
+    engine_args.worker_use_ray = True
+
+    return VLLMDeployment.bind(
+        engine_args=engine_args,
+        response_role="assistant",
+        lora_modules=None,
+        prompt_adapters=None,
+        request_logger=None,
+        chat_template=None,
+    )
+
+
+deployment = build_app(
+    {
+        "model": os.environ["MODEL_ID"],
+        "tensor-parallel-size": os.environ["TENSOR_PARALLELISM"],
+        "pipeline-parallel-size": os.environ["PIPELINE_PARALLELISM"],
+    }
 )
