+++ workloads/training/LLMs/dpo-singlenode/kaiwojob-dpo.yaml
@@ -26,7 +26,7 @@ spec:
     storageEnabled: true
     storageClassName: multinode
     data:
-      storageSize: 10Mi
+      storageSize: 20Mi
       mountPath: /workload
       download:
         git:
