+++ workloads/training/LLMs/full-parameter-pretraining/full-param-zero3-single-multinode/kaiwojob-llama-3.1-8b-instruct.yaml
@@ -38,7 +38,7 @@ spec:
     storageEnabled: true
     storageClassName: multinode
     data:
-      storageSize: 10Mi
+      storageSize: 20Mi
       mountPath: /workload
       download:
         git:
