+++ workloads/training/LLMs/lora-supervised-finetuning/lora-sft-zero3-single-multinode/kaiwojob-llama-3.1-8b-instruct.yaml
@@ -40,7 +40,7 @@ spec:
     storageEnabled: true
     storageClassName: multinode
     data:
-      storageSize: 10Mi
+      storageSize: 20Mi
       mountPath: /workload
       download:
         git:
